Home screen viewmodel integrated

+++ app/src/main/java/com/example/letterboxn/domain/model/ReviewLocalItem.kt
@@ -3,7 +3,7 @@ package com.example.letterboxn.domain.model
 data class ReviewLocalItem (
     val movieId : Int,
     val review : String,
-    val rating : Float,
+    val rating : Double,
     val reviewDate : String,
     val reviewId : Int = 0
 )
